se adaptan algunas cosas a la informacion del cliente

+++ models/informacion_proveedor.go
@@ -4,7 +4,7 @@ package models
 type InformacionProveedor struct {
 	Id                      int                `orm:"column(id_proveedor)"`
 	Tipopersona             string             `orm:"column(tipopersona)"`
-	NumDocumento            int64            `orm:"column(num_documento);pk"`
+	NumDocumento            float64            `orm:"column(num_documento);pk"`
 	IdCiudadContacto        *Ciudad            `orm:"column(id_ciudad_contacto);rel(fk)"`
 	Direccion               string             `orm:"column(direccion)"`
 	Correo                  string             `orm:"column(correo)"`
