+++ models/respuesta.go
@@ -3,7 +3,7 @@ package models
 type Respuesta struct {
 	Id	           int
 	Nombre_Cont         string
-	NumDocumento         int64
+	NumDocumento         float64
 	Valor_bruto string
 	Valor_neto  string
 	Conceptos   *[]ConceptosResumen
